requirements.txt
loguru~=0.7.3
redis~=5.2.1
pydantic~=2.9.2
SQLAlchemy~=2.0.38
fastapi~=0.115.8
PyJWT~=2.10.1
bcrypt~=4.2.1
pytest~=8.3.4
alembic~=1.13.3
aio_pika~=9.5.4